chore: remove react support and refactor codebase
