--- demo/index.tsx
@@ -1,7 +1,0 @@
-import './index.css';
-import { createRoot } from 'react-dom/client';
-import { Demo } from './Demo.js';
-
-const container = document.getElementById('root')!;
-const root = createRoot(container);
-root.render(<Demo />);
