+++ demo/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue';
 import Demo from './Demo.vue';
-import '../index.css';
+import './index.css';
 
 createApp(Demo).mount('#app');
