+++ index.html
@@ -3,98 +3,10 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>use-stick-to-bottom Dual Framework Demo</title>
-    <style>
-      body {
-        margin: 0;
-        padding: 20px;
-        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-          'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-          sans-serif;
-        background: #f8fafc;
-      }
-      .demo-container {
-        max-width: 1400px;
-        margin: 0 auto;
-      }
-      .demo-header {
-        text-align: center;
-        margin-bottom: 2rem;
-      }
-      .framework-demos {
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-        gap: 2rem;
-        margin-bottom: 2rem;
-      }
-      .framework-section {
-        background: white;
-        border-radius: 12px;
-        padding: 1.5rem;
-        box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-      }
-      .framework-title {
-        margin: 0 0 1rem 0;
-        color: #1e293b;
-        text-align: center;
-      }
-      .nav-buttons {
-        display: flex;
-        justify-content: center;
-        gap: 1rem;
-        margin-bottom: 1rem;
-      }
-      .nav-button {
-        padding: 0.5rem 1rem;
-        border: 2px solid #e2e8f0;
-        background: white;
-        color: #64748b;
-        text-decoration: none;
-        border-radius: 8px;
-        font-weight: 500;
-        transition: all 0.2s;
-      }
-      .nav-button:hover {
-        border-color: #3b82f6;
-        color: #3b82f6;
-      }
-      .nav-button.active {
-        background: #3b82f6;
-        border-color: #3b82f6;
-        color: white;
-      }
-      @media (max-width: 768px) {
-        .framework-demos {
-          grid-template-columns: 1fr;
-        }
-      }
-    </style>
+    <title>use-stick-to-bottom Vue 3 Demo</title>
   </head>
   <body>
-    <div class="demo-container">
-      <div class="demo-header">
-        <h1>use-stick-to-bottom</h1>
-        <p>Dual Framework Demo - React & Vue 3</p>
-        <div class="nav-buttons">
-          <a href="/" class="nav-button active">Combined Demo</a>
-          <a href="/demo/react.html" class="nav-button">React Only</a>
-          <a href="/vue.html" class="nav-button">Vue Only</a>
-        </div>
-      </div>
-      
-      <div class="framework-demos">
-        <div class="framework-section">
-          <h2 class="framework-title">⚛️ React Implementation</h2>
-          <div id="react-root"></div>
-        </div>
-        
-        <div class="framework-section">
-          <h2 class="framework-title">💚 Vue 3 Implementation</h2>
-          <div id="vue-root"></div>
-        </div>
-      </div>
-    </div>
-    
-    <script type="module" src="/demo/combined.tsx"></script>
+    <div id="app"></div>
+    <script type="module" src="/demo/main.ts"></script>
   </body>
 </html>
