+++ src/useStickToBottom.ts
@@ -3,40 +3,16 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import {
-  type DependencyList,
-  type MutableRefObject,
-  type RefCallback,
-  useCallback,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
-
-export interface StickToBottomState {
-  scrollTop: number;
-  lastScrollTop?: number;
-  ignoreScrollToTop?: number;
-  targetScrollTop: number;
-  calculatedTargetScrollTop: number;
-  scrollDifference: number;
-  resizeDifference: number;
-
-  animation?: {
-    behavior: 'instant' | Required<SpringAnimation>;
-    ignoreEscapes: boolean;
-    promise: Promise<boolean>;
-  };
-  lastTick?: number;
-  velocity: number;
-  accumulated: number;
-
-  escapedFromLock: boolean;
-  isAtBottom: boolean;
-  isNearBottom: boolean;
-
-  resizeObserver?: ResizeObserver;
-}
+import { computed, onUnmounted, reactive, ref, watch } from 'vue';
+import type {
+  Animation,
+  ScrollToBottom,
+  SpringAnimation,
+  StickToBottomInstance,
+  StickToBottomOptions,
+  StickToBottomState,
+  StopScroll,
+} from './types.js';
 
 const DEFAULT_SPRING_ANIMATION = {
   /**
@@ -63,100 +39,49 @@ const DEFAULT_SPRING_ANIMATION = {
   mass: 1.25,
 };
 
-export interface SpringAnimation
-  extends Partial<typeof DEFAULT_SPRING_ANIMATION> {}
-
-export type Animation = ScrollBehavior | SpringAnimation;
-
-export interface ScrollElements {
-  scrollElement: HTMLElement;
-  contentElement: HTMLElement;
-}
-
-export type GetTargetScrollTop = (
-  targetScrollTop: number,
-  context: ScrollElements
-) => number;
-
-export interface StickToBottomOptions extends SpringAnimation {
-  resize?: Animation;
-  initial?: Animation | boolean;
-  targetScrollTop?: GetTargetScrollTop;
-}
-
-export type ScrollToBottomOptions =
-  | ScrollBehavior
-  | {
-      animation?: Animation;
-
-      /**
-       * Whether to wait for any existing scrolls to finish before
-       * performing this one. Or if a millisecond is passed,
-       * it will wait for that duration before performing the scroll.
-       *
-       * @default false
-       */
-      wait?: boolean | number;
-
-      /**
-       * Whether to prevent the user from escaping the scroll,
-       * by scrolling up with their mouse.
-       */
-      ignoreEscapes?: boolean;
-
-      /**
-       * Only scroll to the bottom if we're already at the bottom.
-       *
-       * @default false
-       */
-      preserveScrollPosition?: boolean;
-
-      /**
-       * The extra duration in ms that this scroll event should persist for.
-       * (in addition to the time that it takes to get to the bottom)
-       *
-       * Not to be confused with the duration of the animation -
-       * for that you should adjust the animation option.
-       *
-       * @default 0
-       */
-      duration?: number | Promise<void>;
-    };
-
-export type ScrollToBottom = (
-  scrollOptions?: ScrollToBottomOptions
-) => Promise<boolean> | boolean;
-export type StopScroll = () => void;
-
 const STICK_TO_BOTTOM_OFFSET_PX = 70;
 const SIXTY_FPS_INTERVAL_MS = 1000 / 60;
 const RETAIN_ANIMATION_DURATION_MS = 350;
 
 let mouseDown = false;
 
-globalThis.document?.addEventListener('mousedown', () => {
-  mouseDown = true;
-});
+// Global mouse tracking for selection detection
+if (typeof globalThis !== 'undefined' && globalThis.document) {
+  globalThis.document.addEventListener('mousedown', () => {
+    mouseDown = true;
+  });
 
-globalThis.document?.addEventListener('mouseup', () => {
-  mouseDown = false;
-});
+  globalThis.document.addEventListener('mouseup', () => {
+    mouseDown = false;
+  });
 
-globalThis.document?.addEventListener('click', () => {
-  mouseDown = false;
-});
+  globalThis.document.addEventListener('click', () => {
+    mouseDown = false;
+  });
+}
 
-export const useStickToBottom = (
+export function useStickToBottom(
   options: StickToBottomOptions = {}
-): StickToBottomInstance => {
-  const [escapedFromLock, updateEscapedFromLock] = useState(false);
-  const [isAtBottom, updateIsAtBottom] = useState(options.initial !== false);
-  const [isNearBottom, setIsNearBottom] = useState(false);
+): StickToBottomInstance {
+  const scrollRef = ref<HTMLElement | null>(null);
+  const contentRef = ref<HTMLElement | null>(null);
 
-  const optionsRef = useRef<StickToBottomOptions>(null!);
-  optionsRef.current = options;
+  const escapedFromLock = ref(false);
+  const isAtBottom = ref(options.initial !== false);
+  const isNearBottom = ref(false);
 
-  const isSelecting = useCallback(() => {
+  const optionsRef = ref<StickToBottomOptions>(options);
+
+  // Keep options up to date
+  watch(
+    () => options,
+    (newOptions) => {
+      optionsRef.value = newOptions;
+    },
+    { deep: true }
+  );
+
+  const isSelecting = (): boolean => {
     if (!mouseDown) {
       return false;
     }
@@ -168,456 +93,414 @@ export const useStickToBottom = (
 
     const range = selection.getRangeAt(0);
     return (
-      range.commonAncestorContainer.contains(scrollRef.current) ||
-      scrollRef.current?.contains(range.commonAncestorContainer)
+      (scrollRef.value
+        ? range.commonAncestorContainer.contains(scrollRef.value)
+        : false) ||
+      (scrollRef.value?.contains(range.commonAncestorContainer) ?? false)
     );
-  }, []);
+  };
 
-  const setIsAtBottom = useCallback((isAtBottom: boolean) => {
-    state.isAtBottom = isAtBottom;
-    updateIsAtBottom(isAtBottom);
-  }, []);
+  const setIsAtBottom = (value: boolean) => {
+    isAtBottom.value = value;
+  };
 
-  const setEscapedFromLock = useCallback((escapedFromLock: boolean) => {
-    state.escapedFromLock = escapedFromLock;
-    updateEscapedFromLock(escapedFromLock);
-  }, []);
+  const setEscapedFromLock = (value: boolean) => {
+    escapedFromLock.value = value;
+  };
 
-  // biome-ignore lint/correctness/useExhaustiveDependencies: not needed
-  const state = useMemo<StickToBottomState>(() => {
-    let lastCalculation:
-      | { targetScrollTop: number; calculatedScrollTop: number }
-      | undefined;
+  const setIsNearBottom = (value: boolean) => {
+    isNearBottom.value = value;
+  };
 
-    return {
-      escapedFromLock,
-      isAtBottom,
-      resizeDifference: 0,
-      accumulated: 0,
-      velocity: 0,
-      listeners: new Set(),
+  const state = reactive<StickToBottomState>({
+    resizeDifference: 0,
+    accumulated: 0,
+    velocity: 0,
+    ignoreScrollToTop: undefined,
+    lastScrollTop: undefined,
+    lastTick: undefined,
+    animation: undefined,
+    resizeObserver: undefined,
+  });
 
-      get scrollTop() {
-        return scrollRef.current?.scrollTop ?? 0;
-      },
-      set scrollTop(scrollTop: number) {
-        if (scrollRef.current) {
-          scrollRef.current.scrollTop = scrollTop;
-          state.ignoreScrollToTop = scrollRef.current.scrollTop;
-        }
-      },
+  // Helper functions instead of getters for better performance
+  const getScrollTop = (): number => {
+    return scrollRef.value?.scrollTop ?? 0;
+  };
 
-      get targetScrollTop() {
-        if (!scrollRef.current || !contentRef.current) {
-          return 0;
-        }
+  const setScrollTop = (scrollTop: number) => {
+    if (scrollRef.value) {
+      scrollRef.value.scrollTop = scrollTop;
+      state.ignoreScrollToTop = scrollRef.value.scrollTop;
+    }
+  };
 
-        return (
-          scrollRef.current.scrollHeight - 1 - scrollRef.current.clientHeight
-        );
-      },
-      get calculatedTargetScrollTop() {
-        if (!scrollRef.current || !contentRef.current) {
-          return 0;
-        }
+  const getTargetScrollTop = (): number => {
+    if (!scrollRef.value || !contentRef.value) {
+      return 0;
+    }
+    return scrollRef.value.scrollHeight - 1 - scrollRef.value.clientHeight;
+  };
 
-        const { targetScrollTop } = this;
+  const getCalculatedTargetScrollTop = (): number => {
+    if (!scrollRef.value || !contentRef.value) {
+      return 0;
+    }
 
-        if (!options.targetScrollTop) {
-          return targetScrollTop;
-        }
+    const targetScrollTop = getTargetScrollTop();
 
-        if (lastCalculation?.targetScrollTop === targetScrollTop) {
-          return lastCalculation.calculatedScrollTop;
-        }
+    if (!optionsRef.value.targetScrollTop) {
+      return targetScrollTop;
+    }
 
-        const calculatedScrollTop = Math.max(
-          Math.min(
-            options.targetScrollTop(targetScrollTop, {
-              scrollElement: scrollRef.current,
-              contentElement: contentRef.current,
-            }),
-            targetScrollTop
-          ),
-          0
-        );
+    return Math.max(
+      Math.min(
+        optionsRef.value.targetScrollTop(targetScrollTop, {
+          scrollElement: scrollRef.value,
+          contentElement: contentRef.value,
+        }),
+        targetScrollTop
+      ),
+      0
+    );
+  };
 
-        lastCalculation = { targetScrollTop, calculatedScrollTop };
+  const getScrollDifference = (): number => {
+    return getCalculatedTargetScrollTop() - getScrollTop();
+  };
 
+  const scrollToBottom: ScrollToBottom = (scrollOptions = {}) => {
+    if (typeof scrollOptions === 'string') {
+      scrollOptions = { animation: scrollOptions };
+    }
+
+    if (!scrollOptions.preserveScrollPosition) {
+      setIsAtBottom(true);
+    }
+
+    const waitElapsed = Date.now() + (Number(scrollOptions.wait) || 0);
+    const behavior = mergeAnimations(optionsRef.value, scrollOptions.animation);
+    const { ignoreEscapes = false } = scrollOptions;
+
+    let durationElapsed: number;
+    let startTarget = getCalculatedTargetScrollTop();
+
+    if (scrollOptions.duration instanceof Promise) {
+      scrollOptions.duration.finally(() => {
+        durationElapsed = Date.now();
+      });
+    } else {
+      durationElapsed = waitElapsed + (scrollOptions.duration ?? 0);
+    }
+
+    const next = async (): Promise<boolean> => {
+      const promise = new Promise<boolean>((resolve) => {
         requestAnimationFrame(() => {
-          lastCalculation = undefined;
-        });
-
-        return calculatedScrollTop;
-      },
-
-      get scrollDifference() {
-        return this.calculatedTargetScrollTop - this.scrollTop;
-      },
-
-      get isNearBottom() {
-        return this.scrollDifference <= STICK_TO_BOTTOM_OFFSET_PX;
-      },
-    };
-  }, []);
-
-  const scrollToBottom = useCallback<ScrollToBottom>(
-    (scrollOptions = {}) => {
-      if (typeof scrollOptions === 'string') {
-        scrollOptions = { animation: scrollOptions };
-      }
-
-      if (!scrollOptions.preserveScrollPosition) {
-        setIsAtBottom(true);
-      }
-
-      const waitElapsed = Date.now() + (Number(scrollOptions.wait) || 0);
-      const behavior = mergeAnimations(
-        optionsRef.current,
-        scrollOptions.animation
-      );
-      const { ignoreEscapes = false } = scrollOptions;
-
-      let durationElapsed: number;
-      let startTarget = state.calculatedTargetScrollTop;
-
-      if (scrollOptions.duration instanceof Promise) {
-        scrollOptions.duration.finally(() => {
-          durationElapsed = Date.now();
-        });
-      } else {
-        durationElapsed = waitElapsed + (scrollOptions.duration ?? 0);
-      }
-
-      const next = async (): Promise<boolean> => {
-        const promise = new Promise(requestAnimationFrame).then(() => {
-          if (!state.isAtBottom) {
+          if (!isAtBottom.value) {
             state.animation = undefined;
-
-            return false;
+            resolve(false);
+            return;
           }
 
-          const { scrollTop } = state;
+          const scrollTop = getScrollTop();
           const tick = performance.now();
           const tickDelta =
             (tick - (state.lastTick ?? tick)) / SIXTY_FPS_INTERVAL_MS;
-          state.animation ||= { behavior, promise, ignoreEscapes };
+
+          state.animation = state.animation || {
+            behavior,
+            promise,
+            ignoreEscapes,
+          };
 
           if (state.animation.behavior === behavior) {
             state.lastTick = tick;
           }
 
           if (isSelecting()) {
-            return next();
+            next().then(resolve);
+            return;
           }
 
           if (waitElapsed > Date.now()) {
-            return next();
+            next().then(resolve);
+            return;
           }
 
           if (
-            scrollTop < Math.min(startTarget, state.calculatedTargetScrollTop)
+            scrollTop < Math.min(startTarget, getCalculatedTargetScrollTop())
           ) {
             if (state.animation?.behavior === behavior) {
               if (behavior === 'instant') {
-                state.scrollTop = state.calculatedTargetScrollTop;
-                return next();
+                setScrollTop(getCalculatedTargetScrollTop());
+                next().then(resolve);
+                return;
               }
 
               state.velocity =
                 (behavior.damping * state.velocity +
-                  behavior.stiffness * state.scrollDifference) /
+                  behavior.stiffness * getScrollDifference()) /
                 behavior.mass;
               state.accumulated += state.velocity * tickDelta;
-              state.scrollTop += state.accumulated;
+              setScrollTop(getScrollTop() + state.accumulated);
 
-              if (state.scrollTop !== scrollTop) {
+              if (getScrollTop() !== scrollTop) {
                 state.accumulated = 0;
               }
             }
 
-            return next();
+            next().then(resolve);
+            return;
           }
 
           if (durationElapsed > Date.now()) {
-            startTarget = state.calculatedTargetScrollTop;
-
-            return next();
+            startTarget = getCalculatedTargetScrollTop();
+            next().then(resolve);
+            return;
           }
 
           state.animation = undefined;
 
-          /**
-           * If we're still below the target, then queue
-           * up another scroll to the bottom with the last
-           * requested animatino.
-           */
-          if (state.scrollTop < state.calculatedTargetScrollTop) {
-            return scrollToBottom({
+          // If we're still below the target, queue up another scroll
+          if (getScrollTop() < getCalculatedTargetScrollTop()) {
+            const result = scrollToBottom({
               animation: mergeAnimations(
-                optionsRef.current,
-                optionsRef.current.resize
+                optionsRef.value,
+                optionsRef.value.resize
               ),
               ignoreEscapes,
               duration: Math.max(0, durationElapsed - Date.now()) || undefined,
             });
+            if (result instanceof Promise) {
+              result.then(resolve);
+            } else {
+              resolve(result);
+            }
+            return;
           }
 
-          return state.isAtBottom;
+          resolve(isAtBottom.value);
+        });
+      });
+
+      promise.then((result) => {
+        requestAnimationFrame(() => {
+          if (!state.animation) {
+            state.lastTick = undefined;
+            state.velocity = 0;
+          }
         });
 
-        return promise.then((isAtBottom) => {
-          requestAnimationFrame(() => {
-            if (!state.animation) {
-              state.lastTick = undefined;
-              state.velocity = 0;
-            }
-          });
+        return result;
+      });
 
-          return isAtBottom;
-        });
-      };
+      return promise;
+    };
 
-      if (scrollOptions.wait !== true) {
-        state.animation = undefined;
-      }
+    if (scrollOptions.wait !== true) {
+      state.animation = undefined;
+    }
 
-      if (state.animation?.behavior === behavior) {
-        return state.animation.promise;
-      }
+    if (state.animation?.behavior === behavior) {
+      return state.animation.promise;
+    }
 
-      return next();
-    },
-    [setIsAtBottom, isSelecting, state]
-  );
+    return next();
+  };
 
-  const stopScroll = useCallback((): void => {
+  const stopScroll: StopScroll = () => {
     setEscapedFromLock(true);
     setIsAtBottom(false);
-  }, [setEscapedFromLock, setIsAtBottom]);
+    state.animation = undefined;
+    state.velocity = 0;
+    state.accumulated = 0;
+  };
 
-  const handleScroll = useCallback(
-    ({ target }: Event) => {
-      if (target !== scrollRef.current) {
-        return;
-      }
-
-      const { scrollTop, ignoreScrollToTop } = state;
-      let { lastScrollTop = scrollTop } = state;
-
-      state.lastScrollTop = scrollTop;
-      state.ignoreScrollToTop = undefined;
-
-      if (ignoreScrollToTop && ignoreScrollToTop > scrollTop) {
-        /**
-         * When the user scrolls up while the animation plays, the `scrollTop` may
-         * not come in separate events; if this happens, to make sure `isScrollingUp`
-         * is correct, set the lastScrollTop to the ignored event.
-         */
-        lastScrollTop = ignoreScrollToTop;
-      }
-
-      setIsNearBottom(state.isNearBottom);
-
-      /**
-       * Scroll events may come before a ResizeObserver event,
-       * so in order to ignore resize events correctly we use a
-       * timeout.
-       *
-       * @see https://github.com/WICG/resize-observer/issues/25#issuecomment-248757228
-       */
-      setTimeout(() => {
-        /**
-         * When theres a resize difference ignore the resize event.
-         */
-        if (state.resizeDifference || scrollTop === ignoreScrollToTop) {
-          return;
-        }
-
-        if (isSelecting()) {
-          setEscapedFromLock(true);
-          setIsAtBottom(false);
-          return;
-        }
-
-        const isScrollingDown = scrollTop > lastScrollTop;
-        const isScrollingUp = scrollTop < lastScrollTop;
-
-        if (state.animation?.ignoreEscapes) {
-          state.scrollTop = lastScrollTop;
-          return;
-        }
-
-        if (isScrollingUp) {
-          setEscapedFromLock(true);
-          setIsAtBottom(false);
-        }
-
-        if (isScrollingDown) {
-          setEscapedFromLock(false);
-        }
-
-        if (!state.escapedFromLock && state.isNearBottom) {
-          setIsAtBottom(true);
-        }
-      }, 1);
-    },
-    [setEscapedFromLock, setIsAtBottom, isSelecting, state]
-  );
-
-  const handleWheel = useCallback(
-    ({ target, deltaY }: WheelEvent) => {
-      let element = target as HTMLElement;
-
-      while (!['scroll', 'auto'].includes(getComputedStyle(element).overflow)) {
-        if (!element.parentElement) {
-          return;
-        }
-
-        element = element.parentElement;
-      }
-
-      /**
-       * The browser may cancel the scrolling from the mouse wheel
-       * if we update it from the animation in meantime.
-       * To prevent this, always escape when the wheel is scrolled up.
-       */
-      if (
-        element === scrollRef.current &&
-        deltaY < 0 &&
-        scrollRef.current.scrollHeight > scrollRef.current.clientHeight &&
-        !state.animation?.ignoreEscapes
-      ) {
-        setEscapedFromLock(true);
-        setIsAtBottom(false);
-      }
-    },
-    [setEscapedFromLock, setIsAtBottom, state]
-  );
-
-  const scrollRef = useRefCallback((scroll) => {
-    scrollRef.current?.removeEventListener('scroll', handleScroll);
-    scrollRef.current?.removeEventListener('wheel', handleWheel);
-    scroll?.addEventListener('scroll', handleScroll, { passive: true });
-    scroll?.addEventListener('wheel', handleWheel, { passive: true });
-  }, []);
-
-  const contentRef = useRefCallback((content) => {
-    state.resizeObserver?.disconnect();
-
-    if (!content) {
+  const handleScroll = (event: Event) => {
+    if (event.target !== scrollRef.value) {
       return;
     }
 
-    let previousHeight: number | undefined;
+    const scrollTop = getScrollTop();
+    const ignoreScrollToTop = state.ignoreScrollToTop;
+    let lastScrollTop = state.lastScrollTop ?? scrollTop;
 
-    state.resizeObserver = new ResizeObserver(([entry]) => {
-      const { height } = entry.contentRect;
-      const difference = height - (previousHeight ?? height);
+    state.lastScrollTop = scrollTop;
+    state.ignoreScrollToTop = undefined;
 
-      state.resizeDifference = difference;
+    if (ignoreScrollToTop && ignoreScrollToTop > scrollTop) {
+      lastScrollTop = ignoreScrollToTop;
+    }
 
-      /**
-       * Sometimes the browser can overscroll past the target,
-       * so check for this and adjust appropriately.
-       */
-      if (state.scrollTop > state.targetScrollTop) {
-        state.scrollTop = state.targetScrollTop;
+    setIsNearBottom(getScrollDifference() <= STICK_TO_BOTTOM_OFFSET_PX);
+
+    // Use setTimeout to handle potential race conditions with ResizeObserver
+    setTimeout(() => {
+      if (state.resizeDifference || scrollTop === ignoreScrollToTop) {
+        return;
       }
 
-      setIsNearBottom(state.isNearBottom);
+      if (isSelecting()) {
+        setEscapedFromLock(true);
+        setIsAtBottom(false);
+        return;
+      }
 
-      if (difference >= 0) {
-        /**
-         * If it's a positive resize, scroll to the bottom when
-         * we're already at the bottom.
-         */
-        const animation = mergeAnimations(
-          optionsRef.current,
-          previousHeight
-            ? optionsRef.current.resize
-            : optionsRef.current.initial
-        );
+      const isScrollingDown = scrollTop > lastScrollTop;
+      const isScrollingUp = scrollTop < lastScrollTop;
 
-        scrollToBottom({
-          animation,
-          wait: true,
-          preserveScrollPosition: true,
-          duration:
-            animation === 'instant' ? undefined : RETAIN_ANIMATION_DURATION_MS,
+      if (state.animation?.ignoreEscapes) {
+        setScrollTop(lastScrollTop);
+        return;
+      }
+
+      if (isScrollingUp) {
+        setEscapedFromLock(true);
+        setIsAtBottom(false);
+      }
+
+      if (isScrollingDown) {
+        setEscapedFromLock(false);
+      }
+
+      if (
+        !escapedFromLock.value &&
+        getScrollDifference() <= STICK_TO_BOTTOM_OFFSET_PX
+      ) {
+        setIsAtBottom(true);
+      }
+    }, 1);
+  };
+
+  const handleWheel = (event: WheelEvent) => {
+    let element = event.target as HTMLElement;
+
+    while (!['scroll', 'auto'].includes(getComputedStyle(element).overflow)) {
+      if (!element.parentElement) {
+        return;
+      }
+      element = element.parentElement;
+    }
+
+    // Prevent scroll escape when wheel scrolled up
+    if (
+      element === scrollRef.value &&
+      event.deltaY < 0 &&
+      scrollRef.value.scrollHeight > scrollRef.value.clientHeight &&
+      !state.animation?.ignoreEscapes
+    ) {
+      setEscapedFromLock(true);
+      setIsAtBottom(false);
+    }
+  };
+
+  // Setup event listeners and ResizeObserver
+  watch(
+    scrollRef,
+    (newScrollRef, oldScrollRef) => {
+      // Remove old listeners
+      oldScrollRef?.removeEventListener('scroll', handleScroll);
+      oldScrollRef?.removeEventListener('wheel', handleWheel);
+
+      // Add new listeners
+      if (newScrollRef) {
+        newScrollRef.addEventListener('scroll', handleScroll, {
+          passive: true,
         });
-      } else {
-        /**
-         * Else if it's a negative resize, check if we're near the bottom
-         * if we are want to un-escape from the lock, because the resize
-         * could have caused the container to be at the bottom.
-         */
-        if (state.isNearBottom) {
-          setEscapedFromLock(false);
-          setIsAtBottom(true);
-        }
+        newScrollRef.addEventListener('wheel', handleWheel, { passive: true });
+      }
+    },
+    { immediate: true }
+  );
+
+  watch(
+    contentRef,
+    (newContentRef) => {
+      state.resizeObserver?.disconnect();
+
+      if (!newContentRef) {
+        return;
       }
 
-      previousHeight = height;
+      let previousHeight: number | undefined;
 
-      /**
-       * Reset the resize difference after the scroll event
-       * has fired. Requires a rAF to wait for the scroll event,
-       * and a setTimeout to wait for the other timeout we have in
-       * resizeObserver in case the scroll event happens after the
-       * resize event.
-       */
-      requestAnimationFrame(() => {
-        setTimeout(() => {
-          if (state.resizeDifference === difference) {
-            state.resizeDifference = 0;
+      state.resizeObserver = new ResizeObserver(([entry]) => {
+        const { height } = entry.contentRect;
+        const difference = height - (previousHeight ?? height);
+
+        state.resizeDifference = difference;
+
+        // Check for overscroll and adjust
+        if (getScrollTop() > getTargetScrollTop()) {
+          setScrollTop(getTargetScrollTop());
+        }
+
+        setIsNearBottom(getScrollDifference() <= STICK_TO_BOTTOM_OFFSET_PX);
+
+        if (difference >= 0) {
+          // Positive resize: scroll to bottom if already at bottom
+          const animation = mergeAnimations(
+            optionsRef.value,
+            previousHeight ? optionsRef.value.resize : optionsRef.value.initial
+          );
+
+          scrollToBottom({
+            animation,
+            wait: true,
+            preserveScrollPosition: true,
+            duration:
+              animation === 'instant'
+                ? undefined
+                : RETAIN_ANIMATION_DURATION_MS,
+          });
+        } else {
+          // Negative resize: check if we're near bottom
+          if (getScrollDifference() <= STICK_TO_BOTTOM_OFFSET_PX) {
+            setEscapedFromLock(false);
+            setIsAtBottom(true);
           }
-        }, 1);
-      });
-    });
+        }
 
-    state.resizeObserver?.observe(content);
-  }, []);
+        previousHeight = height;
+
+        // Reset resize difference after scroll event
+        requestAnimationFrame(() => {
+          setTimeout(() => {
+            if (state.resizeDifference === difference) {
+              state.resizeDifference = 0;
+            }
+          }, 1);
+        });
+      });
+
+      state.resizeObserver.observe(newContentRef);
+    },
+    { immediate: true }
+  );
+
+  onUnmounted(() => {
+    state.resizeObserver?.disconnect();
+    state.resizeObserver = undefined;
+    state.animation = undefined;
+    state.velocity = 0;
+    state.accumulated = 0;
+    scrollRef.value?.removeEventListener('scroll', handleScroll);
+    scrollRef.value?.removeEventListener('wheel', handleWheel);
+  });
 
   return {
     contentRef,
     scrollRef,
     scrollToBottom,
     stopScroll,
-    isAtBottom: isAtBottom || isNearBottom,
+    isAtBottom: computed(() => isAtBottom.value || isNearBottom.value),
     isNearBottom,
     escapedFromLock,
     state,
   };
-};
-
-export interface StickToBottomInstance {
-  contentRef: React.MutableRefObject<HTMLElement | null> &
-    React.RefCallback<HTMLElement>;
-  scrollRef: React.MutableRefObject<HTMLElement | null> &
-    React.RefCallback<HTMLElement>;
-  scrollToBottom: ScrollToBottom;
-  stopScroll: StopScroll;
-  isAtBottom: boolean;
-  isNearBottom: boolean;
-  escapedFromLock: boolean;
-  state: StickToBottomState;
 }
 
-function useRefCallback<T extends (ref: HTMLElement | null) => any>(
-  callback: T,
-  deps: DependencyList
-) {
-  // biome-ignore lint/correctness/useExhaustiveDependencies: not needed
-  const result = useCallback((ref: HTMLElement | null) => {
-    result.current = ref;
-    return callback(ref);
-  }, deps) as any as MutableRefObject<HTMLElement | null> &
-    RefCallback<HTMLElement>;
-
-  return result;
-}
-
+// Animation merging utility
 const animationCache = new Map<string, Readonly<Required<SpringAnimation>>>();
 
 function mergeAnimations(...animations: (Animation | boolean | undefined)[]) {
@@ -630,7 +513,7 @@ function mergeAnimations(...animations: (Animation | boolean | undefined)[]) {
       continue;
     }
 
-    if (typeof animation !== 'object') {
+    if (typeof animation !== 'object' || animation === null) {
       continue;
     }
 
@@ -649,3 +532,13 @@ function mergeAnimations(...animations: (Animation | boolean | undefined)[]) {
 
   return instant ? 'instant' : animationCache.get(key)!;
 }
+
+export type {
+  GetTargetScrollTop,
+  ScrollToBottom,
+  ScrollToBottomOptions,
+  StickToBottomInstance,
+  StickToBottomOptions,
+  StickToBottomState,
+  StopScroll,
+} from './types.js';
