+++ tailwind.config.js
@@ -2,7 +2,7 @@ const { iconsPlugin } = require('@egoist/tailwindcss-icons');
 
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: ['./index.html', './demo/**/*.{ts,tsx}'],
+  content: ['./index.html', './demo/**/*.{vue,ts}'],
   theme: {
     extend: {},
   },
