+++ tsconfig.app.json
@@ -12,7 +12,6 @@
     "isolatedModules": true,
     "moduleDetection": "force",
     "declaration": true,
-    "jsx": "react-jsx",
 
     /* Linting */
     "strict": true,
