--- vite.config.react.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-export default defineConfig({
-  plugins: [react()],
-  root: './',
-  build: {
-    rollupOptions: {
-      input: './index.html',
-    },
-  },
-});
