--- vite.config.vue.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'vite';
-import vue from '@vitejs/plugin-vue';
-
-export default defineConfig({
-  plugins: [vue()],
-  root: './',
-  build: {
-    rollupOptions: {
-      input: './vue.html',
-    },
-  },
-});
