--- vue.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>use-stick-to-bottom Vue demo</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="/demo/vue/main.ts"></script>
-  </body>
-</html>
